gpu: HACK: Disable main_i2c4
Disables main_i2c4 to use it as an interrupt for a guest GPU. Since the
main GPU interrupt is already taken by root-cell in the jailhouse
configuration, a workaround was made to forward the interrupt to the
guests via the i2c4. This will be reverted once CLEC based routing is
enabled.

Signed-off-by: Erick Narvaez <e-narvaez@ti.com>
Signed-off-by: Gowtham Tammana <g-tammana@ti.com>