drm/tidss: cleanup dma related settings
Clean up and improve DMA related settings. We didn't set many of the
relevant registers, like fifo_thresholds at all.

The specific values used are "sounds nice" guesses. We need to somehow
measure the DMA performance and improve these to cover most of the use
cases, but this patch should serve as a good starting point.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>