drm/tidss: static partition of wb
DSS7 has support for partitioning the write-back pipeline among the
several compute entities and they can be independently driven from
different cores.

This patch adds support for statically partitioning the write-back
pipeline using device-tree nodes.

If there is a DT node that specifies how the write-back pipeline should
be used, then the information is used to populate WB accordingly.

The .has_writeback flag is only check directly to decide if we need to
check the DT data for related writeback nodes.
dispc7_has_writeback() from this point forward will only check if WB is
managed by tidss or not.

Signed-off-by: Benoit Parrot <bparrot@ti.com>