ARM: dts: dra7-evm: Add CMA pool for GPU
The CMA reserved node has been added for GPU core and enabled for the
DRA7 EVM board. The GPU device cannot access LPAE addresses, and as
such the CMA region is set in the *highmem32* region. This enables the
driver to allocate from the CMA region and there by not limiting all
GPU allocations to come only from normal zone.

Signed-off-by: Gowtham Tammana <g-tammana@ti.com>