ARM: dts: dra76-evm: Add CMA pool for GPU device
The CMA reserved memory node is added for the GPU device on the DRA76
EVM board, and assigned to respective node. The configuration in here
matches to that on the DRA7 EVM board. This enables the driver to
allocate from CMA region and there by not limiting all GPU allocations
to come only from normal zone.

Signed-off-by: Gowtham Tammana <g-tammana@ti.com>