drm/omap: fix scaling limits
"drm/omap: dynamically assign hw overlays to planes" added a call to
drm_atomic_helper_check_plane_state() in omap_plane_atomic_check(), with
scaling limits that were supposed to allow all the valid setups.

However, the limits were backwards, and the limit meant for upscaling
limited downscaling, and vice versa.

Also, the limit for downscaling is a bit too restrictive (1/4).

Fix the limits so that they allow down to 1/8 downscaling and up to x8
upscaling.

Note that the "real" scaling limits are more complex and calculated in
dispc. Unfortunately that happens only at commit phase, but that is
another topic.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reported-by: Ram Prasad <x0038811@ti.com>
Reviewed-by: Benoit Parrot <bparrot@ti.com>