drm/bridge: cdns-mhdp: Protect firmware mailbox messaging with mutex
The communication with the DisplayPort IP firmware happens over a
simple mailbox. Any concurrency will get it out of sync, so it is
better to protect all transactions with a mutex. This patch is needed
if there is any access to the mailbox from the IRQ handler. Note that
a mutex can not be used in an interrupt context, but this is not a
problem since cdns-mhdp is using a threaded interrupt handler.

Signed-off-by: Jyri Sarha <jsarha@ti.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ti.com>