arm64: dts: ti: k3-j721: move DP routing and pinmux to board DT file
We should not setup DP routes in the SoC dt file, but in the board dt
file. Nor should we set the pinctrl in SoC dt, especially if the
property is referring to a setting that is found in the board DT file.

Fix this by removing the ports and pinctrl settings from mhdp node in
k3-j721e-main.dtsi, and moving these to the board dt files. Also
remove obsolete TODO pinmux comment.

Signed-off-by: Jyri Sarha <jsarha@ti.com>