arm64: dts: ti: k3-j721: cleanup DP connector data
Clean up the DP connector data by fixing bad indent, remove the comment,
and adding connector label.

There is no binding for connectors yet, so this data is not used.
However, in mainline a connector binding has been sent for review, and
while it doesn't yet have DP support, this binding has been made to be
compatible in style with the proposed one.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>