PCI: endpoint: pci-epf-ntb: Set error status on command failure
pci-epf-ntb supports 3 commands: command to configure doorbell,
command to configure memory window and command to indicate link up.
These commands are given to the endpoint by the host. All of these
commands can fail in the endpoint. However there is no mechanism for
the endpoint to indicate the failure to the host. Add a new bit in the
STATUS register to indicate any error in command processing to the host.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>