drm/omap: fix dma_addr refcounting
"drm/omap: use refcount API to track the number of users of dma_addr"
changed omap_gem.c to use refcounting API to track dma_addr uses.
However, the driver only tracks the refcounts for non-contiguous
buffers, and the patch didn't fully take this in account.

After the patch, the driver always decreased refcount in omap_gem_unpin,
instead of decreasing the refcount only for non-contiguous buffers. This
leads to refcounting mismatch.

As for the non-contiguous cases the refcount is never increased, fix
this issue by returning from omap_gem_unpin if the buffer being unpinned
is not non-contiguous.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>