drm/tidss: bail out from atomic_flush if there's no event
We expect crtc->state->event to be set when tidss_crtc_atomic_flush is
called, and we have a WARN_ON() to check it. However, if
crtc->state->event is NULL, we'll still continue, and later again check
if crtc->state->event is set.

Clean this up by returning from the function if the WARN_ON() is hit. As
this ensures crtc->state->event is not NULL, we can drop the latter
check.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reviewed-by: Jyri Sarha <jsarha@ti.com>