TEMP: ARM: dts: dra7-ipu-common: Limit IPU early boot to only IPU1
The dra7-ipu-common-early-boot.dtsi file has the necessary changes to
configure both the IPU1 and IPU2 remote processors for 'late attach' mode.
Remove the changes for IPU2 for the moment to limit the 'late attach' mode
for IPU1 only. This is being done due to the limitations with some of the
current IPU2 firmware files.

Signed-off-by: Suman Anna <s-anna@ti.com>