ti_config_fragments: v8_rpmsg: Enable OMAP HwSpinlock driver
The OMAP HwSpinlock driver is reused for supporting the HwSpinlock IP
within K3 platforms. So, enable the fully upstreamed OMAP HwSpinlock
driver in the v8 rpmsg config file. The HwSpinlock core was already
enabled by default in the ARM v8 defconfig file. There are currently
no direct users of the HwSpinlock IP, but the built-in is the preferred
choice for this driver.

Signed-off-by: Suman Anna <s-anna@ti.com>