ti_config_fragments: rpmsg: Enable OMAP HwSpinlock and Mailbox drivers
Enable the fully upstreamed (on OMAP2+ SoCs) OMAP HwSpinlock and
and OMAP Mailbox drivers including the corresponding HwSpinlock and
Mailbox cores in the base IPC/RPMsg domain config fragment file for
v7 platforms. The options are added based on leveraging the ARM
multi_v7_defconfig file. These are chosen to be built-in as they
are the preferred choice for these drivers.

Signed-off-by: Suman Anna <s-anna@ti.com>