drm/tidss: increase preload to avoid underflows
AM6 EVM shows underflows with some downscaling use cases. One way to
decrease the possibility of underflows is to adjust the preload value.
The driver currently sets it to a rather low value.

This patch makes the HW use the fifo high-threshold value as the preload
value, which should minimize the possibility of underflows from preload
perspective.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>