drm/tidss: remove underflow handling
It was found out that we may have lots of underflows even in a normal,
working, setup.

This may happen during the DSS preload period, where DSS fills all its
fifos and buffers at maximum speed. Underflow indicates that the DMA
cannot keep up with the DSS's internal pixel processing, but as the
pixels are not consumed yet by anything, there's no harm in these
underflows. It effectively just indicates that DSS's internal processing
is being limited by the DMA.

If an underflow happens during the active period of the display, it
causes a sync-lost interrupt as the videoport cannot get the pixels it
needs. This sync-lost is often caused by an underflow, though.

So, underflow alone is not an indication of a problem, but sync lost
always is.

This patch removes underflow handling from the driver to get rid of the
underflow error print floods. In the future we may want to improve this
so that via a debug option some kind of underflow statistics could be
collected.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reviewed-by: Jyri Sarha <jsarha@ti.com>