ntb: ntb_hw_epf: Do not set dma_mask for NTB device (struct ntb_dev)
NTB device is not a real device and the piece of hardware actually
doing the DMA is the PCI device itself. As a result the DMA
capability of the NTB device is not configured (DMA ops of NTB device
is not populated) and dma_coerce_mask_and_coherent() on NTB device
fails.

Remove configuring dma_mask with dma_coerce_mask_and_coherent() on
NTB device and fix ntb_perf.c and ntb_tool.c to use PCI device for
dma_alloc_coherent(). ntb_transport.c already used PCI device for
dma_alloc_coherent().

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>