ti_config_fragments: v8_rpmsg: Enable rpmsg client sample
Enable the in-kernel rpmsg client sample module so that it can be
used as the unit-test for verifying rpmsg communication with the
various remote processors present on AM65x and J721E SoCs.

The rpmsg communication can be verified using firmwares from the
PDK IPC for both AM65x & J721E SoCs; and also from IPC 3.x on
AM65x SoCs.

Signed-off-by: Suman Anna <s-anna@ti.com>