net: ethernet: ti: icssg-prueth: Add multicast filter support
Each ICSSG port supports 8 MAC filters (FT1) that can be used
to implement multicast filtering. However, since 8 filters are not
sufficient for a real world case, we optimize it a bit by reserving
first 2 filters for usually required multicast addresses.

  1) 01-80-C2-00-00-XX Known Service Ethernet Multicast addresses
  2) 01-00-5e-00-00-XX Local Network Control Block
                       (224.0.0.0 - 224.0.0.255  (224.0.0/24))

Then we use the remaining 6 FT1 filters as a pool for an other
multicast addresses.

At any instant if we run out of the filters then we just turn on all
multicast packets.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>