net: ethernet: icssg: introduce a generic f/w command interface
Refactor the command creation code to a separate function,
emac_send_command() to prepare for additional command support
in the driver. Adjust the error messages in emac_send_command()
to include the command type. Use constant for shutdown command
type instead of using hardcoded value and add a comment
for the shutdown command in icssg_config.h.

Acked-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>