net: ethernet: icssg-prueth: implement f/w command to set speed/duplex settings
As per firmware configuration interface specification, Firmware expects
driver to send auto negotiated speed and duplex settings to its management
interface as soon as the link is up. This patch implements the required
pstate speed/duplex command. As the command re-use the RGMII CFG status
bits for speed and duplex parameters in the command, read the same from
the register using helper functions to create the command. While at it,
move the spin_{lock/unlock} functions in emac_adjust_link() closer to
the register update functions so as to avoid holding double lock.

Acked-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>