dma-buf: Move locks to only protect framework data
The DMA-BUF framework mutex lock is used to protect the attachments list,
but in some spots it also surrounds the exporter callback attach. This is
not needed and causes a lockdep warning when exporters call back into
the DMA-BUF framework. Move the locking to only guard the framework
data it is meant to protect.

Signed-off-by: Andrew F. Davis <afd@ti.com>