ARM: dts: AM33xx-l4: Add PRUSS sysc node
Add the DT nodes for the PRU-ICSS on AM33xx family of SoCs. The
AM33xx SoCs contain a single PRU-ICSS instance and is represented
by a pruss node and other child nodes. PRU-ICSS is not supported
on AM3352 SoC though in the AM33xx family, so the nodes are added
under the corresponding disabled interconnect target module node
in the common am33xx-l4 dtsi file. The target module node should
be enabled in only those derivative board files that use a SoC
containing PRU-ICSS.

The PRUSS subsystem node contains the entire address space. The
various sub-modules of the PRU-ICSS are represented as individual
child nodes (so platform devices themselves) of the PRUSS subsystem
node. These include the two PRU cores and the interrupt controller.
The Industrial Ethernet Peripheral (IEP), the Real Time Media
Independent Interface controller (MII_RT), and the CFG sub-module
are represented as syscon nodes. All the Data RAMs are represented
within a child node of its own named 'memories' without any
compatible.

The DT nodes use all standard properties. The regs property in
the PRU nodes define the addresses for the Instruction RAM, the
Debug and Control sub-modules for that PRU core. The firmware for
each PRU core is defined through a 'firmware-name' property.

The default names for the firmware images for each PRU core are
defined as follows (these can be adjusted either in derivative
board dts files or through sysfs at runtime if required):
     PRU-ICSS PRU0 Core: am335x-pru1_0-fw
     PRU-ICSS PRU1 Core: am335x-pru1_1-fw

Signed-off-by: Suman Anna <s-anna@ti.com>