ARM: dts: AM4372: Add the PRU-ICSS0 DT node
The AM437x SoC has a second smaller PRU-ICSS subsystem (PRUSS0)
in addition to the primary PRUSS1 instance. The PRUSS0 has less
DRAM per PRU, and no Shared DRAM among other minor differences.
The IEP and MII_RT modules even though present within the IP are
not pinned out.

This PRUSS0 instance has a weird SoC integration. It shares the
same L3 OCP interconnect interface with PRUSS1, and also shares
its reset line and clocks. Any external accesses from PRUSS0
requires the PRUSS1's PRUSS_SYSCFG register to be programmed
properly. That said, it is its own IP instance (a cut-down version),
and so it has been added as an independent node (sibling node to
PRUSS1 node) and a child node of the corresponding PRUSS target
module interconnect node. This allows the PRUSS0 instance to be
enabled/disabled independently of the PRUSS1 instance.

The nodes are added under the corresponding disabled interconnect
target module node in the common am4372 dtsi file, as not every
SoC in the AM437x family has the PRU-ICSSs (eg: AM4372 SoC lacks
the support). The target module node should be enabled in only
those derivative board files that use a SoC containing the PRU-ICSS.
The individual PRUSS node can be disabled in the corresponding
board dts file if desired.

The default names for the firmware images for each PRU core are
defined as follows (these can be adjusted either in derivative
board dts files or through sysfs at runtime if required):
     PRU-ICSS0 PRU0 Core: am437x-pru0_0-fw
     PRU-ICSS0 PRU1 Core: am437x-pru0_1-fw

Signed-off-by: Suman Anna <s-anna@ti.com>