TEMP: soc: ti: pruss: support CORECLK_MUX and IEPCLK_MUX
The IEPCLK_MUX is present on all SoCs whereas the CORECLK_MUX
is present only on AM65x SoCs. Add support for both these CLK
muxes.

This allows the clock rates and clock parents for these
to be controlled through DT leveraging the clk infrastructure
for configuring the default parents and rates.

NOTE:
Support is limited only for K3 AM65x SoCs at the moment as DT
needs to be fixed up for other SoCs.

Signed-off-by: Roger Quadros <rogerq@ti.com>
[s-anna@ti.com: regmap name fix, few cleanups and 5.4 port]
Signed-off-by: Suman Anna <s-anna@ti.com>