ARM: dts: am57xx-idk-common: Add PRUSS2 Ethernet application node
Add PRUSS Ethernet application node for PRU-ICSS2. We don't add this on
PRU-ICSS1 as it is not connected to the SoC by default on AM57xx IDKs.

On AM572x-idk and AM574x-idk, the PHYs have individual GPIO RESET
lines so keep them in PHY nodes to reflect actual hardware setup.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>