net: netcp: fix error in link status display
When Ethernet PHY is present, the Ethernet link status is determined
by the combined statuses of the PHY link and the SGMII link. That is,
Ethernet link is down if either status is down, and Ethernet link is
up if both statuses are up.

Prior to this patch, when Ethernet link state changes, the PHY link
status is displayed, regardless whether the combined status is up or
down. However, when the Ethernet link goes down, due a race condition,
the PHY link state may not yet correctly reflect the down state while
the SGMII link state does. Thus blindly displaying the PHY link state
is incorrect when the Ethernet link goes down.

This patch fixes the problem by displaying the combined state when
Ethernet link goes down and if PHY is present. When Ethernet link goes
up, the PHY link state is displayed if PHY is present. When PHY is absent,
the SGMII link status is displayed in any link state change.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>