phy: keystone: serdes driver for NetCP 1GBE and PCIE
On TI's Keystone platforms, such as the GBE ethernet switch, and
PCIe controller require the use of a SERDES for converting SoC
parallel data into serialized data that can be output over a
high-speed electrical interface, and also converting high-speed
serial input data into parallel data that can be processed by the
SoC.  The SerDeses used by those peripherals, though they may be
different, are largely similar in functionality and setup.

This patch provides a SERDES PHY driver implementation that can be
used by the above mentioned peripheral drivers to configure their
respective SERDES hardware.

The SERDES configurations are divided mainly in two parts, namely,
a common configuration which is done regardless a SERDES lane is
actually used or not, and a lane specific configuration which is
done when a lane is actually used by, for example, a network interface.

The common configuration is performed by the SERDES provider
driver, ie., phy-keystone-serdes.c, at driver probe. The lane
specific configuration is invoked by the corresponding peripheral
driver, for example network switch driver, through the PHY API.

The SERDES IP is owned by a 3rd party and this patch uses no comments
in the driver as recommended by the vendor. Also the license header
used in the file is in accordance with the vendor's recommendation
of using BSD license for the driver file.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Jack Manbeck <j-manbeck2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>