net: keystone: update netcp ethss driver to use common SERDES PHY driver
Prior to this patch, 1GBE inherits the SERDES configurations from
u-boot.  This patch updates the Keystone ethernet driver to use the
generic Keystone SERDES PHY driver for SERDES initialization
and configuration.  The common SERDES PHY driver supports peripherals
on Keystone platforms that require SERDES namely 1GBE & PCIE.

Additionally, this patch introduced serdes lane recovery procedure for
1GBE links.  When the remote device that the 1GBE device is connected
to is restarted for whatever reason, the receive direction of a 1GBE
SERDES lane will need to be reset in order to sync up again with
the remote device's transmit pattern. When 1GBE driver detects a SERDES
lane down, it will start a periodic attempt to recover the receive
direction of that lane. The recovery is done when MAC_MAC_ENFORCED
interface configuration is used.

This patch also adopts the remodeling of one serdes lane as a separate
PHY device. Hence phy_init() is invoked per each Ethernet interface.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Rex Chang <rchang@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>