serial: add pru software uart (pru_suart) driver
This adds a new serial driver that supports the PRU software based UART.
With the current version of the PRU UART firmware (v1.0), up to three UART
ports are supported per PRU core. Each character is two bytes in the FIFO.
Hardware flow control is supported if CTS/RTS pins are defined, but
software flow control is not.

Cc: Jason Reeder <jreeder@ti.com>
Acked-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Bin Liu <b-liu@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>