dt-bindings: remoteproc: omap: Update ti,autosuspend-delay-ms description
The OMAP remoteproc binding has been updated to add some more details
for the "ti,autosuspend-delay-ms" property that is used for the default
runtime pm autosuspend delay in Linux. This is an optional property
with a default value of 10000 ms, but the runtime auto-suspend can be
disabled for a specific remote processor by default by using a negative
value.

Signed-off-by: Suman Anna <s-anna@ti.com>